Engine:  - ported iOS -> Mac,  - got e-mail opening working,  - added ChibiXM+HappyHTTP to include path,  - got font rendering working in Mac.

+++ UiLepra/Source/Mac/UiMacInput.mm
@@ -62,7 +62,7 @@ void MacInputElement::SetValue(int pValue)
 
 	if (GetType() == DIGITAL)
 	{
-		Parent::SetValue(pValue);
+		Parent::SetValue((float)pValue);
 	}
 	/*else if (GetInterpretation() == RELATIVE_AXIS)
 	{
